Added basic navbar and created products page
