+++ src/app/layout.tsx
@@ -3,6 +3,8 @@ import type { Metadata } from "next"
 import { Roboto } from "next/font/google"
 import { type ReactNode } from "react"
 
+import { Navbar } from "@/ui/organisms/Navbar"
+
 const robotoFont = Roboto({ weight: "400", subsets: ["latin"] })
 
 export const metadata: Metadata = {
@@ -13,7 +15,10 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: ReactNode }) {
 	return (
 		<html lang="en">
-			<body className={robotoFont.className}>{children}</body>
+			<body className={robotoFont.className}>
+				<Navbar />
+				{children}
+			</body>
 		</html>
 	)
 }
